Build wheel on CI
