+++ .gitignore
@@ -11,11 +11,8 @@ jdbc/.gradle
 jdbc/build
 odbc_tests/cmake-build
 parameters.json
-pep249_dbapi/build
-pep249_dbapi/pep249_dbapi.egg-info/
-pep249_dbapi/reports
-pep249_dbapi/snowflake_pep249_dbapi.egg-info/
 **/rsa_key.p8
+.DS_Store
 
 # Python/Cython build artifacts
 python/src/snowflake/ud_connector/_internal/nanoarrow_cpp/ArrowIterator/arrow_stream_iterator.cpp
