+++ python/src/snowflake/ud_connector/_core/.gitignore
@@ -1,0 +1,3 @@
+*
+!.gitignore
+!__init__.py
